refactor: assignColor 유틸함수를 이용하여 컬러를 할당하도록 변경 (#102)
