+++ src/pages/user/Schedule/ScheduleMain/DetailScheduleWrapper/DetailSchedule/index.tsx
@@ -15,6 +15,7 @@ import {
 } from '../../../../../../slices/schedule/scheduleSlice';
 import type { RootState } from '../../../../../../state/store';
 import { useSchedule } from '../../../../../../hooks/useSchedule';
+import { assignColor } from '../../../../../../utils';
 
 interface DetailSchedulProps {
   formattedClickedDate: string;
@@ -132,21 +133,6 @@ const DetailSchedule = ({
     }
 
     return { top, height };
-  };
-
-  const assignColor = (number: number, type: string) => {
-    const colorSaturation = type === 'background' ? 's95' : 's60';
-    const color = [
-      colors.scale.secondary[colorSaturation],
-      colors.scale.tertiary[colorSaturation],
-      colors.scale.neutral[colorSaturation],
-      colors.scale.primary[colorSaturation],
-      colors.scale.danger[colorSaturation],
-    ];
-
-    const assingedColor = color[number % 5];
-
-    return assingedColor;
   };
 
   const getTargetSchedule = (
